# jpa-teaoclock/src/main/java/com/dabramov/teaoclock/entity/Addressee.java
package com.dabramov.teaoclock.entity;

import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
import lombok.Data;

import javax.persistence.*;
import java.io.Serializable;

@Data
@Entity
@Table(name = "ADDRESSEE")
@JsonIgnoreProperties("hibernateLazyInitializer")
public class Addressee implements Serializable {
    @Id
    @GeneratedValue(strategy = GenerationType.AUTO)
    @Column(name = "ID")
    private Long id;

    @Column(name = "EMAIL", nullable = false, unique = true)
    private String email;

    @ManyToOne(fetch = FetchType.LAZY, cascade = {CascadeType.MERGE, CascadeType.PERSIST})
    @JoinColumn(name = "MESSAGE_ID")
    @JsonIgnoreProperties("addressees")
    private Message message;
}