# jpa-teaoclock/src/main/resources/swagger-contract.json
{"swagger":"2.0","info":{"description":"Api Documentation","version":"1.0","title":"Api Documentation","termsOfService":"urn:tos","contact":{},"license":{"name":"Apache 2.0","url":"http://www.apache.org/licenses/LICENSE-2.0"}},"host":"localhost:8081","basePath":"/","tags":[{"name":"addressee-controller","description":"Addressee Controller"},{"name":"message-controller","description":"Message Controller"}],"paths":{"/api/addressees":{"get":{"tags":["addressee-controller"],"summary":"getAddressees","operationId":"getAddresseesUsingGET","produces":["*/*"],"parameters":[{"name":"ids","in":"query","description":"ids","required":false,"type":"array","items":{"type":"integer","format":"int64"},"collectionFormat":"multi"}],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/Addressee"}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}},"post":{"tags":["addressee-controller"],"summary":"saveAddressee","operationId":"saveAddresseeUsingPOST","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"addressee","description":"addressee","required":true,"schema":{"$ref":"#/definitions/Addressee"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Addressee"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/api/addressees/{id}":{"get":{"tags":["addressee-controller"],"summary":"getAllAddresses","operationId":"getAllAddressesUsingGET","produces":["*/*"],"parameters":[{"name":"id","in":"path","description":"id","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Addressee"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}},"delete":{"tags":["addressee-controller"],"summary":"deleteAddressee","operationId":"deleteAddresseeUsingDELETE","produces":["*/*"],"parameters":[{"name":"id","in":"path","description":"id","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"OK"},"204":{"description":"No Content"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"}}}},"/api/messages":{"get":{"tags":["message-controller"],"summary":"getMessages","operationId":"getMessagesUsingGET","produces":["*/*"],"parameters":[{"name":"ids","in":"query","description":"ids","required":false,"type":"array","items":{"type":"integer","format":"int64"},"collectionFormat":"multi"}],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/Message"}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}},"post":{"tags":["message-controller"],"summary":"saveMessage","operationId":"saveMessageUsingPOST","consumes":["application/json"],"produces":["*/*"],"parameters":[{"name":"addressees[0].email","in":"query","required":false,"type":"string"},{"name":"addressees[0].id","in":"query","required":false,"type":"integer","format":"int64"},{"name":"addressees[0].message.content","in":"query","required":false,"type":"string"},{"name":"addressees[0].message.id","in":"query","required":false,"type":"integer","format":"int64"},{"name":"addressees[0].message.sender","in":"query","required":false,"type":"string"},{"name":"addressees[0].message.subject","in":"query","required":false,"type":"string"},{"name":"content","in":"query","required":false,"type":"string"},{"name":"id","in":"query","required":false,"type":"integer","format":"int64"},{"name":"sender","in":"query","required":false,"type":"string"},{"name":"subject","in":"query","required":false,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Message"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/api/messages/{id}":{"get":{"tags":["message-controller"],"summary":"getMessage","operationId":"getMessageUsingGET","produces":["*/*"],"parameters":[{"name":"id","in":"path","description":"id","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Message"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}},"delete":{"tags":["message-controller"],"summary":"deleteMessage","operationId":"deleteMessageUsingDELETE","produces":["*/*"],"parameters":[{"name":"id","in":"path","description":"id","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"OK"},"204":{"description":"No Content"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"}}}}},"definitions":{"Addressee":{"type":"object","properties":{"email":{"type":"string"},"id":{"type":"integer","format":"int64"},"message":{"$ref":"#/definitions/Message"}},"title":"Addressee"},"Message":{"type":"object","properties":{"addressees":{"type":"array","items":{"$ref":"#/definitions/Addressee"}},"content":{"type":"string"},"id":{"type":"integer","format":"int64"},"sender":{"type":"string"},"subject":{"type":"string"}},"title":"Message"}}}